:sparkles: Elpi plugin skeleton for the Sub command
The only thing that must change in the Demo.v file is that terms that are not constants must be inside parentheses. In this case, the integers in which substitutions are made (4, 256). If we do not add parentheses, they can still be parsed, but as Elpi integers, which we do not want here.